fix: complicated issue in game-service client/server

+++ protocol/src/main/kotlin/me/ihdeveloper/humans/service/protocol/Packet.kt
@@ -16,6 +16,6 @@ abstract class Packet(
     }
 
     protected fun writeNonce(source: PacketBuffer, nonce: Int) {
-        source.writeInt(nonce)
+        source.writeShort(nonce)
     }
 }
